Fix - config
